examen/requirements.txt
kafka-python-ng>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
pandas>=2.2.0
pyspark==3.5.0
streamlit>=1.31.0
plotly>=5.18.0
python-json-logger>=2.0.7

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Optional
python-dotenv>=1.0.0
